refactor: update inspector code to be self-standing
