+++ inspector/app/renderer/components/Session/CapabilityControl.js
@@ -3,7 +3,7 @@ import { Switch, Input } from 'antd';
 import SessionStyles from './Session.css';
 import { remote } from 'electron';
 import { FileOutlined } from '@ant-design/icons';
-import { INPUT } from '../AntdTypes';
+import { INPUT } from '../../../../../shared/components/AntdTypes';
 import _ from 'lodash';
 import log from 'electron-log';
