+++ inspector/app/renderer/components/Session/ServerTabPcloudy.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Form, Row, Col, Input } from 'antd';
 import SessionStyles from './Session.css';
-import { INPUT } from '../AntdTypes';
+import { INPUT } from '../../../../../shared/components/AntdTypes';
 
 const FormItem = Form.Item;
