+++ inspector/app/renderer/components/Session/ServerTabTestobject.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { Form, Input, Radio, Row, Col } from 'antd';
 import SessionStyles from './Session.css';
-import { INPUT } from '../AntdTypes';
+import { INPUT } from '../../../../../shared/components/AntdTypes';
 
 const FormItem = Form.Item;
