+++ inspector/app/renderer/components/Session/Session.js
@@ -12,7 +12,7 @@ import SessionStyles from './Session.css';
 import CloudProviders from './CloudProviders';
 import CloudProviderSelector from './CloudProviderSelector';
 import { LikeOutlined } from '@ant-design/icons';
-import { BUTTON } from '../AntdTypes';
+import { BUTTON } from '../../../../../shared/components/AntdTypes';
 
 const {TabPane} = Tabs;
