+++ inspector/package.json
@@ -148,6 +148,7 @@
     "@babel/core": "^7.4.0",
     "@babel/plugin-transform-runtime": "^7.4.0",
     "@babel/register": "^7.4.0",
+    "appium-fake-driver": "^1.0.1",
     "asyncbox": "^2.5.3",
     "babel-eslint": "^10.0.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
