+++ server/package.json
@@ -153,7 +153,6 @@
     "@babel/core": "^7.4.0",
     "@babel/plugin-transform-runtime": "^7.4.0",
     "@babel/register": "^7.4.0",
-    "appium-fake-driver": "^1.0.0",
     "asyncbox": "^2.5.3",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "chai": "^4.2.0",
